feat: Change the typing for the AIConfig (#16)
This commit inlines the `AIConfigData` class into the existing
`AIConfig`.

It also introduces a new type, `ModelConfig` to replace the much looser
dictionary config that was originally in play. This new model contains
specific, typed properties for the model id, temperature, and max
tokens. Additional model-specific attributes can be accessed as well.