fix: Distinguish AIConfig from DefaultAIConfig
Consumers of this SDK are responsible for providing some default value
to the `config` method. If that type is of `AIConfig`, customers have
the option of providing a tracker instance that will never be useful.

To help discourage this, we are making two separate types -- one to
represent the default ai config, and one to represent the actual
selected config.