fix: Return AIConfig and LDAITracker separately (#23)
Consumers of this SDK are responsible for providing some default value
to the `config` method. If that type is of `AIConfig`, customers have
the option of providing a tracker instance that will never be useful.

To address this, we are going to remove the tracker from the `AIConfig`
type entirely. Instead, we are going to return the `AIConfig` and the
`LDAITracker` separately.